# build/swagger.yaml
swagger: '2.0'
basePath: 'localhost:3000'
host: 'localhost:3000'
servers:
  - description: Base Url
    url: 'localhost:3000'
info:
  version: 1.0.0
  title: ''
paths:
  /users:
    get:
      tags:
        - User
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
  /user:
    post:
      requestBody:
        description: Body
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                loginType:
                  type: string
      tags:
        - User
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
  '/user/{userId}':
    get:
      parameters:
        - in: path
          name: userId
          schema:
            type: number
          required: true
          description: 조회할 유저의 ID
      tags:
        - User
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
    update:
      tags:
        - User
      parameter: null
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
    delete:
      parameters:
        - in: path
          name: userId
          schema:
            type: number
          required: true
          description: 삭제할 유저의 ID
      tags:
        - User
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
  /articles:
    get:
      tags:
        - Articles
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
  /article:
    post:
      requestBody:
        description: Body
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                loginType:
                  type: string
      tags:
        - Articles
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
  '/article/{articleId}':
    get:
      parameters:
        - in: path
          name: articleId
          schema:
            type: number
          required: true
          description: 조회할 글의 ID
      tags:
        - Articles
      responses:
        '200':
          '400':
            description: BAD_REQUEST
          '401':
            description: FAILED_TO_ATHORIZE
          '404':
            description: NOT_FOUND
          '500':
            description: SERVER_ERROR
          description: OK
    update:
      tags:
        - Articles
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
    delete:
      tags:
        - Articles
      parameters:
        - in: path
          name: articleId
          schema:
            type: number
          required: true
          description: 삭제할 글의 ID
      responses:
        '200':
          description: OK
        '400':
          description: BAD_REQUEST
        '401':
          description: FAILED_TO_ATHORIZE
        '404':
          description: NOT_FOUND
        '500':
          description: SERVER_ERROR
definitions:
  UserDto:
    type: object
    required:
      - nickname
      - login_type
    properties:
      id:
        type: number
      loginType:
        type: string
        enum:
          - NONE
          - KAKAO
          - GOOGLE
          - APPLE
      externalId:
        type: number
      nickname:
        type: string
      gender:
        type: string
      email:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      profileImg:
        type: number
      ageRange:
        type: string
      birthday:
        type: string
      createDate:
        type: Date
      modifyDate:
        type: Date
  ArticleDto:
    type: object
    required:
      - creatorId
      - title
      - contents
    properties:
      id:
        type: number
      creatorId:
        type: number
      categoryId:
        type: number
      title:
        type: string
      viewCount:
        type: number
      likeCount:
        type: number
      tags:
        type: string
      contents:
        type: string
      isPublished:
        type: boolean
      createDate:
        type: Date
      modifyDate:
        type: Date